// .cursor/recover-540858c/cms/src/types/api.ts
export interface LiveRevenueLogItem {
    id: string
    revenueType: number
    revenueTypeText?: string
    roomId: string
    liveRecordId: string
    senderId: string
    senderNickname?: string
    senderAvatar?: string
    receiverId: string
    receiverNickname?: string
    receiverAvatar?: string
    bizId: string
    bizName?: string
    count: number
    unitPrice: number
    totalAmount: number
    status?: number
    statusText?: string
    createdAt?: string | null
}